refactor: reorganize frontend folder structure
