+++ frontend/app/components/home-page/EventCard.vue
@@ -25,7 +25,7 @@
 </template>
 
 <script setup lang="ts">
-import type { Event } from '../types'
+  import type { Event } from '~/types/home'
 
 interface Props {
   event: Event
