+++ frontend/app/components/home-page/EventCarousel.vue
@@ -1,33 +1,22 @@
 <template>
   <section 
-    class="mb-5 position-relative"
+    class="position-relative"
     @touchstart="handleTouchStart"
     @touchmove="handleTouchMove"
     @touchend="handleTouchEnd"
   >
     <div class="row g-3">
       <div
-        v-for="(event, idx) in visibleSlides"
-        :key="event.id"
-        :class="isMobile ? 'col-12' : 'col-md-6'"
+        v-for="item in visibleItems"
+        :key="item.id"
+        :class="colClass"
       >
-        <div class="event-card-large position-relative rounded-4 overflow-hidden">
-          <img
-            :src="event.imageUrl"
-            :alt="event.title"
-            class="w-100 h-100 object-fit-cover"
-          />
-          <button
-            class="btn btn-light position-absolute bottom-0 start-0 m-3 rounded-2"
-          >
-            Xem chi tiết
-          </button>
-        </div>
+        <slot :item="item" />
       </div>
     </div>
 
     <button
-      v-if="totalSlides > 1 && !isMobile"
+      v-if="canGoPrev && !isMobile"
       class="carousel-nav-btn prev"
       @click="goPrev"
       aria-label="Previous"
@@ -36,7 +25,7 @@
     </button>
 
     <button
-      v-if="totalSlides > 1 && !isMobile"
+      v-if="canGoNext && !isMobile"
       class="carousel-nav-btn next"
       @click="goNext"
       aria-label="Next"
@@ -44,45 +33,54 @@
       &gt;
     </button>
 
-    <div v-if="totalSlides > 1" class="carousel-indicators-dots">
+    <div
+      v-if="showDots"
+      class="carousel-indicators-dots"
+    >
       <button
-        v-for="(_, index) in totalSlides"
+        v-for="(_, index) in totalPages"
         :key="index"
         class="dot"
-        :class="{ active: currentSlide === index }"
-        @click="goToSlide(index)"
-        :aria-label="`Go to slide ${index + 1}`"
+        :class="{ active: currentIndex === index }"
+        @click="goToPage(index)"
+        :aria-label="`Go to page ${index + 1}`"
       />
     </div>
   </section>
 </template>
 
 <script setup lang="ts">
-import { ref, computed, onMounted, onUnmounted } from 'vue'
-import type { Event } from '../types'
+import { ref, onMounted, onUnmounted } from 'vue'
+import { useCarousel } from '~/composables/useCarousel'
 
 interface Props {
-  events: Event[]
+  items: any[]
+  itemsPerPage?: number
+  colClass?: string
+  showDots?: boolean
 }
 
-const props = defineProps<Props>()
-const currentSlide = ref(0)
-const isMobile = ref(false)
+const props = withDefaults(defineProps<Props>(), {
+  itemsPerPage: 4,
+  colClass: 'col-lg-3 col-md-4 col-sm-6',
+  showDots: false,
+})
 
+const isMobile = ref(false)
 const touchStartX = ref(0)
 const touchEndX = ref(0)
 const minSwipeDistance = 50
 
-const itemsPerSlide = computed(() => isMobile.value ? 1 : 2)
-
-const totalSlides = computed(() =>
-  Math.ceil(props.events.length / itemsPerSlide.value)
-)
-
-const visibleSlides = computed(() => {
-  const start = currentSlide.value * itemsPerSlide.value
-  return props.events.slice(start, start + itemsPerSlide.value)
-})
+const {
+  currentIndex,
+  totalPages,
+  canGoPrev,
+  canGoNext,
+  visibleItems,
+  goNext,
+  goPrev,
+  goToPage,
+} = useCarousel(props.items, props.itemsPerPage)
 
 const updateViewport = () => {
   if (typeof window !== 'undefined') {
@@ -107,29 +105,11 @@ const handleTouchEnd = () => {
   const isLeftSwipe = distance > minSwipeDistance
   const isRightSwipe = distance < -minSwipeDistance
 
-  if (isLeftSwipe) {
+  if (isLeftSwipe && canGoNext.value) {
     goNext()
-  } else if (isRightSwipe) {
+  } else if (isRightSwipe && canGoPrev.value) {
     goPrev()
   }
-}
-
-const goPrev = () => {
-  currentSlide.value =
-    currentSlide.value === 0
-      ? totalSlides.value - 1
-      : currentSlide.value - 1
-}
-
-const goNext = () => {
-  currentSlide.value =
-    currentSlide.value === totalSlides.value - 1
-      ? 0
-      : currentSlide.value + 1
-}
-
-const goToSlide = (index: number) => {
-  currentSlide.value = index
 }
 
 onMounted(() => {
@@ -147,17 +127,6 @@ onUnmounted(() => {
 <style scoped>
 section {
   position: relative;
-}
-
-.event-card-large {
-  width: 100%;
-  aspect-ratio: 807 / 460;
-}
-
-@media (max-width: 767.98px) {
-  .event-card-large {
-    aspect-ratio: 4 / 3;
-  }
 }
 
 .carousel-nav-btn {
@@ -191,7 +160,7 @@ section {
 
 .carousel-indicators-dots {
   position: absolute;
-  bottom: -25px;
+  bottom: 15px;
   left: 50%;
   transform: translateX(-50%);
   display: flex;
@@ -203,7 +172,7 @@ section {
   width: 15px;
   height: 15px;
   border-radius: 50%;
-  background-color: var(--bg-reactive-gray);
+  background-color: rgba(255, 255, 255, 0.5);
   cursor: pointer;
   transition: background-color 0.3s;
   border: none;
@@ -211,10 +180,10 @@ section {
 }
 
 .carousel-indicators-dots .dot:hover {
-  background-color: var(--bg-reactive-gray-hover);
+  background-color: rgba(255, 255, 255, 0.8);
 }
 
 .carousel-indicators-dots .dot.active {
-  background-color: #07B3DF;
+  background-color: rgba(255, 255, 255, 1);
 }
 </style>
