+++ frontend/app/components/home-page/SpecialEventCard.vue
@@ -9,7 +9,7 @@
 </template>
 
 <script setup lang="ts">
-import type { Event } from '../types'
+import type { Event } from '~/types/home'
 
 interface Props {
   event: Event
