+++ frontend/app/composables/useCarousel.ts
