+++ frontend/app/composables/useEventFilter.ts
