+++ frontend/app/types/home.d.ts
